@@ -1,21 +1,9 @@
-// The Vue build version to load with the `import` command
-// (runtime-only or standalone) has been set in webpack.base.conf with an alias.
-import Vue from 'vue'
-import App from './App'
+import { createApp } from 'vue'
+import App from './App.vue'
 import router from './router'
 
-import VueCookies from 'vue-cookies'
-import Buefy from 'buefy'
-import 'buefy/dist/buefy.css'
+import './assets/main.css'
 
-Vue.use(Buefy)
-Vue.use(VueCookies)
-Vue.config.productionTip = false
-
-/* eslint-disable no-new */
-new Vue({
-  el: '#app',
-  router,
-  components: { App },
-  template: '<App/>'
-})
+const app = createApp(App)
+app.use(router)
+app.mount('#app')
