@@ -58,7 +58,7 @@
 </template>
 
 <script>
-import HeaderBar from './HeaderBar'
+import HeaderBar from './HeaderBar.vue'
 
 export default {
   name: 'Splash',
