@@ -36,7 +36,7 @@
 </template>
 
 <script>
-import HeaderBar from '../HeaderBar'
+import HeaderBar from '../HeaderBar.vue'
 
 export default {
   name: 'production',
